Make `toString 0` a bunch faster
It's a bit of a dirty trick, but essentially checking a gazillion times
whether that parameter is 0 or not in order to indent correctly takes a
fairly significant chunk of time on huge output.

When I say huge, I'm talking "Hey let's spit out 2MB of html" and this
kind of performance actually matters. So yeah, just passing around an
"indenter" and setting that to `identity` for the `0` case is a bunch
faster.

I have an idea of building the final result up back to front and I think
this could also be a bunch faster, but I'd have to scrap pretty prenting
and that doesn't really seem worth it. On the other hand, though, I
could track the current depth so the point is moot. Cool. Who knew.

Maybe I should write commit messages like this more often, turns out
it's actually useful!